Make GovUkTypography available outside Umbraco. Keep TinyMCE-specific code inside Umbraco.
