+++ GovUk.Frontend.AspNetCore.Extensions/Typography/BackgroundType.cs
@@ -1,4 +1,4 @@
-﻿namespace GovUk.Frontend.Umbraco.Typography
+﻿namespace GovUk.Frontend.AspNetCore.Extensions.Typography
 {
     public enum BackgroundType
     {
