+++ GovUk.Frontend.AspNetCore.Extensions/Typography/TypographyOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace GovUk.Frontend.AspNetCore.Extensions.Typography
+{
+    public class TypographyOptions
+    {
+        public BackgroundType BackgroundType { get; set; } = BackgroundType.Light;
+    }
+}
