+++ GovUk.Frontend.Umbraco/PropertyEditors/ValueFormatters/GovUkTypographyPropertyValueFormatter.cs
@@ -1,5 +1,4 @@
-﻿using GovUk.Frontend.Umbraco.Typography;
-using ThePensionsRegulator.Umbraco;
+﻿using ThePensionsRegulator.Umbraco;
 using ThePensionsRegulator.Umbraco.PropertyEditors;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Models.PublishedContent;
@@ -10,7 +9,7 @@ namespace GovUk.Frontend.Umbraco.PropertyEditors.ValueFormatters
     /// <summary>
     /// Apply GOV.UK typography classes to HTML from the Umbraco rich text editor
     /// </summary>
-    public class GovUkTypographyPropertyValueFormatter : IPropertyValueFormatter
+    public class GovUkTypographyPropertyValueFormatter : TinyMCEPropertyValueFormatterBase, IPropertyValueFormatter
     {
         /// <inheritdoc />
         public bool IsFormatter(IPublishedPropertyType propertyType) => Constants.PropertyEditors.Aliases.TinyMce.Equals(propertyType.EditorAlias);
@@ -20,6 +19,6 @@ namespace GovUk.Frontend.Umbraco.PropertyEditors.ValueFormatters
         /// This property type should return <see cref="IHtmlEncodedString"/> but accept <c>string</c> as well so that
         /// it is possible to provide a string of HTML to <see cref="OverridablePublishedElement.OverrideValue(string, object)"/>.
         /// </remarks>
-        public object FormatValue(object value) => new HtmlEncodedString(GovUkTypography.Apply(value is IHtmlEncodedString html ? html.ToHtmlString() : value as string));
+        public object FormatValue(object value) => ApplyGovUkTypographyToTinyMCE(value);
     }
 }
