--- GovUk.Frontend.Umbraco/Typography/TypographyOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace GovUk.Frontend.Umbraco.Typography
-{
-    public class TypographyOptions
-    {
-        public BackgroundType BackgroundType { get; set; } = BackgroundType.Light;
-        public bool RemoveWrappingParagraphs { get; set; }
-        public bool RemoveWrappingParagraph { get; set; }
-    }
-}
