+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkButton.cshtml
@@ -1,6 +1,5 @@
 ﻿@inherits Umbraco.Cms.Web.Common.Views.UmbracoViewPage<OverridableBlockListItem>;
 @addTagHelper *, GovUk.Frontend.AspNetCore
-@using GovUk.Frontend.Umbraco.Typography
 @using GovUk.Frontend.Umbraco.Models
 @using GovUk.Frontend.Umbraco;
 @using Microsoft.AspNetCore.Mvc.ModelBinding;
