+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkDateInput.cshtml
@@ -3,7 +3,6 @@
 @addTagHelper *, GovUk.Frontend.AspNetCore.Extensions
 @using GovUk.Frontend.Umbraco
 @using GovUk.Frontend.Umbraco.Models
-@using GovUk.Frontend.Umbraco.Typography
 @using GovUk.Frontend.Umbraco.Validation
 @using Microsoft.AspNetCore.Mvc.ModelBinding
 @using ThePensionsRegulator.Umbraco.BlockLists;
