+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkSummaryCard.cshtml
@@ -1,6 +1,5 @@
 ﻿@using GovUk.Frontend.Umbraco;
 @using GovUk.Frontend.Umbraco.Models;
-@using GovUk.Frontend.Umbraco.Typography;
 @using ThePensionsRegulator.Umbraco.BlockLists;
 @using Umbraco.Cms.Core.Models;
 @using Umbraco.Cms.Core.Strings;
