+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkSummaryList.cshtml
@@ -1,7 +1,6 @@
 ﻿@inherits Umbraco.Cms.Web.Common.Views.UmbracoViewPage<OverridableBlockListItem>;
 @addTagHelper *, GovUk.Frontend.AspNetCore
 @using GovUk.Frontend.Umbraco
-@using GovUk.Frontend.Umbraco.Typography
 @using GovUk.Frontend.Umbraco.Models
 @using ThePensionsRegulator.Umbraco.BlockLists;
 @using Umbraco.Cms.Core.Models
