+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkTaskList.cshtml
@@ -2,7 +2,6 @@
 @addTagHelper *, GovUk.Frontend.AspNetCore.Extensions
 @using GovUk.Frontend.AspNetCore.Extensions.HtmlGeneration;
 @using GovUk.Frontend.AspNetCore.Extensions;
-@using GovUk.Frontend.Umbraco.Typography
 @using GovUk.Frontend.Umbraco.Models
 @using GovUk.Frontend.Umbraco;
 @using ThePensionsRegulator.Umbraco.BlockLists;
