+++ GovUk.Frontend.Umbraco/Views/Partials/GOVUK/GovUkTextarea.cshtml
@@ -4,7 +4,6 @@
 @inject IModelPropertyResolver propertyResolver
 @using GovUk.Frontend.AspNetCore.Extensions.Validation
 @using GovUk.Frontend.Umbraco
-@using GovUk.Frontend.Umbraco.Typography
 @using GovUk.Frontend.Umbraco.Models
 @using System.Reflection
 @using System.ComponentModel.DataAnnotations
