+++ ThePensionsRegulator.Frontend.Umbraco/Views/Partials/TPR/TPRBackToMenu.cshtml
@@ -1,7 +1,6 @@
 ﻿@inherits Umbraco.Cms.Web.Common.Views.UmbracoViewPage<OverridableBlockListItem>;
 @addTagHelper *, ThePensionsRegulator.Frontend
 @using GovUk.Frontend.AspNetCore.Extensions.HtmlGeneration;
-@using GovUk.Frontend.Umbraco.Typography
 @using GovUk.Frontend.Umbraco.Models
 @using GovUk.Frontend.Umbraco;
 @using ThePensionsRegulator.Frontend.Umbraco.Services;
